Cleanup util code
